Face-Recognize-api-main/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.4
opencv-python==4.8.1.78
opencv-contrib-python==4.7.0.68
Pillow==10.0.1
deepface==0.0.75
rembg==2.0.53
face-recognition==1.3.0
dlib==19.24.2
imutils
opencv-python-headless
requests

